+++ packages/generator-common/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-cloud-sdk/generator-common
 
+## 2.9.0
+
+### Patch Changes
+
+- Updated dependencies [4c51d3dc]
+  - @sap-cloud-sdk/util@2.9.0
+
 ## 2.8.0
 
 ### Minor Changes